plugins/cd4ml/inference/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
scikit-learn==1.6.1
pandas==2.2.3
nltk==3.9.1
dvc[s3]==3.59.1
dagshub==0.1.0
numpy==2.2.4
joblib==1.4.2
python-jose==3.4.0
pydantic==2.11.3
python-multipart==0.0.20

